requirements.txt
# Core data science libraries
jupyter>=1.0.0
pandas>=1.5.0,<2.3.0
numpy>=1.21.0,<2.0.0

# Visualization libraries
matplotlib>=3.5.0,<3.10.0
seaborn>=0.13.0,<0.14.0
plotly>=5.0.0,<6.0.0

# Machine learning libraries
scikit-learn>=1.3.0,<1.8.0
shap>=0.48.0,<0.49.0
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0

# Web dashboard
streamlit>=1.46.0,<1.47.0

# Image processing
Pillow>=8.0.0,<11.0.0

# Utility libraries
watchdog>=6.0.0,<7.0.0
ipywidgets>=7.6.0,<9.0.0
tqdm>=4.64.0,<5.0.0

# Optional: For enhanced notebook experience
notebook>=6.4.0,<8.0.0
jupyterlab>=3.0.0,<5.0.0